Use explicit branch name

+++ .ci/phpbench-regression-test.sh
@@ -1,9 +1,14 @@
 #!/usr/bin/env bash
 set -e
 
+if [ -z "$TRAVIS_PULL_REQUEST_BRANCH" ]; then
+    echo "PR is not a pull request, skipping benchmarks"
+    exit 0;
+fi
+
 git reset --hard
 git remote add upstream https://github.com/phpactor/worse-reflection
-git fetch upstream master
+git fetch upstream master $TRAVIS_PULL_REQUEST_BRANCH
 
 echo -e "\n\n"
 echo -e "Benchmarking master branch"
@@ -16,7 +21,7 @@ vendor/bin/phpbench run --report=aggregate_compact --progress=dots --retry-thres
 echo -e "\n\n"
 echo -e "Benchmarking current branch and comparing to master"
 echo -e "===================================================\n\n"
-git checkout -
+git checkout $TRAVIS_PULL_REQUEST_BRANCH
 mv composer.lock.pr composer.lock
 composer install --quiet
 vendor/bin/phpbench run --report=aggregate_compact --progress=dots --retry-threshold=2 --uuid=tag:master
